requirements.txt
flask
flask-cors
google-generativeai
pillow
sentence-transformers
faiss-cpu
numpy
torch
transformers
